mes_web/requirements.txt
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0
websockets>=12,<16
paho-mqtt>=1.6,<3.0
openpyxl>=3.1,<4.0
